+++ packages/accordion/CHANGELOG.md
@@ -1,3 +1,6 @@
+## [2.0.1] 2019-09-26 [Release Notes](https://creditandfinance.atlassian.net/wiki/spaces/HUB/pages/803930391/Upcoming+Major+Changes)
+- Bumped lendi-ui dependencies to latest versions
+
 ## [2.0.0] 2019-9-23 2019-8-21 [Release notes](https://creditandfinance.atlassian.net/wiki/spaces/HUB/pages/803930391/Upcoming+Major+Changes)
 - Update typescript to version 3.5.2 and react to version 16.8.4