+++ packages/date-picker/CHANGELOG.md
@@ -1,3 +1,6 @@
+## [4.0.1] 2019-09-26 [Release Notes](https://creditandfinance.atlassian.net/wiki/spaces/HUB/pages/803930391/Upcoming+Major+Changes)
+- Bumped lendi-ui dependencies to latest versions
+
 ## [4.0.0] 2019-9-23 - [HUB-151](https://creditandfinance.atlassian.net/browse/HUB-151)
 
 ## [3.0.0] 2019-9-23 - [HUB-151](https://creditandfinance.atlassian.net/browse/HUB-151)