+++ packages/lender-logos/CHANGELOG.md
@@ -1,4 +1,6 @@
 # Changelog - @lendi-ui/partner-logo
+## [2.0.1] 2019-09-26 [Release Notes](https://creditandfinance.atlassian.net/wiki/spaces/HUB/pages/803930391/Upcoming+Major+Changes)
+- Bumped lendi-ui dependencies to latest versions
 
 ## [2.0.0] 2019-8-8 [${Ticket Number}](https://creditandfinance.atlassian.net/browse.${Ticket Number})