+++ packages/logo/CHANGELOG.md
@@ -1,3 +1,6 @@
+## [6.0.1] 2019-09-26 [Release Notes](https://creditandfinance.atlassian.net/wiki/spaces/HUB/pages/803930391/Upcoming+Major+Changes)
+- Bumped lendi-ui dependencies to latest versions
+
 ## [6.0.0] 2019-9-23 - 2019-08-19 - [HUB-151](https://creditandfinance.atlassian.net/browse/HUB-151)
 ### Updated
 - Update typescript to version 3.5.2 and react to version 16.8.4